requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.14
Django==4.1
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.7
Pillow==9.2.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
pyOpenSSL==22.0.0
pyotp==2.9.0
pyparsing==3.1.1
pytz==2022.2.1
requests==2.31.0
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==2.1.0
vonage==3.11.1
vonage-jwt==1.1.0
wrapt==1.16.0
zope.interface==5.4.0
